+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `Configurable` for reading/writing classes from a `config`, or `json`/`yaml` file
 
 ## [0.1.3] - 2021-11-19
 ### Added